// packages/core/tests/fixtures/lists/zenstack/schema.ts
//////////////////////////////////////////////////////////////////////////////////////////////
// DO NOT MODIFY THIS FILE                                                                  //
// This file is automatically generated by ZenStack CLI and should not be manually updated. //
//////////////////////////////////////////////////////////////////////////////////////////////

/* eslint-disable */

import { type SchemaDef, type AttributeApplication, type FieldDefault, ExpressionUtils } from "@zenstackhq/schema";
export class SchemaType implements SchemaDef {
    provider = {
        type: "postgresql"
    } as const;
    models = {
        Gallery: {
            name: "Gallery",
            fields: {
                id: {
                    name: "id",
                    type: "String",
                    id: true,
                    attributes: [{ name: "@id" }, { name: "@default", args: [{ name: "value", value: ExpressionUtils.call("cuid") }] }] as readonly AttributeApplication[],
                    default: ExpressionUtils.call("cuid") as FieldDefault
                },
                title: {
                    name: "title",
                    type: "String"
                },
                photos: {
                    name: "photos",
                    type: "String",
                    array: true,
                    attributes: [{ name: "@file", args: [{ name: "accept", value: ExpressionUtils.array("String", [ExpressionUtils.literal("image/*")]) }, { name: "cleanup", value: ExpressionUtils.literal(true) }] }] as readonly AttributeApplication[]
                }
            },
            attributes: [
                { name: "@@allow", args: [{ name: "operation", value: ExpressionUtils.literal("all") }, { name: "condition", value: ExpressionUtils.literal(true) }] }
            ] as readonly AttributeApplication[],
            idFields: ["id"],
            uniqueFields: {
                id: { type: "String" }
            }
        }
    } as const;
    plugins = {};
}
export const schema = new SchemaType();